hptune/trainer/requirements.txt
kfp
pandas
numpy
joblib
db-dtypes
pandas-gbq
scikit-learn
google-cloud-bigquery
google-cloud-storage
google-cloud-aiplatform